pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anomaly-detection"
version = "0.1.0"
description = "B2B customer behavioral anomaly detection"
requires-python = ">=3.11"
dependencies = [
    "pandas",
    "numpy",
    "scikit-learn",
    "pyod",
    "mlflow",
    "matplotlib",
    "seaborn",
    "pyyaml",
    "openpyxl",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
